py_tool/pyproject.toml
[project]
name = "py-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.2",
    "kafka-python>=2.0.2",
    "loguru>=0.7.3",
    "pika>=1.3.2",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
]
